requirements.txt
flask==2.0.1
werkzeug==2.0.3
pycryptodome==3.10.1
